requirements.txt
numpy==1.19.4
openpyxl==3.0.5
yahoofinancials==1.6
PyMsgBox==1.0.9
yfinance==0.1.55
requests==2.25.1
pandas==1.2.0
beautifulsoup4==4.9.3
python_dateutil==2.8.1
